requirements.txt
Django >3.2,<3.3
asgiref==3.7.2

django-debug-toolbar
django-extensions
django-robots
django-taggit
django-summernote
django-simple-captcha
django-multi-captcha-admin

Pillow
pytz
sqlparse
tzdata
